TS: WebStream: add optional AsyncIterator methods, and add related casting tests
For informational purposes.
To use actually the asyncIterable interface (with for-await) when available,
users will need to manually cast a `DOM.ReadableStream` to
`ReadableStream` or `PonyfilledReadableStream`.

Also, add tests to ensure the casting works out of the box.
`test/test.tsconfig.json` has been renamed to `test/tsconfig.json` so that TS IDE
integrations automatically apply the corresponding config to the test folder.