TS: add `transform` and `transformAsync` type definitions (#68)
A downside of this solution is that we cannot enforce output type consistency
for the two callbacks since we need parametric inference of the respective
function types.
However, using such callbacks will now result in a union output type, that should signal
the usage might be off.

Also, add tests for legacy type definitions.